Improve code structure: subdirectories, type placement, and naming #181
