+++ src/lib/game/credits-config.ts
@@ -17,3 +17,13 @@ export function make_credits_scroll_bounds(line_count: number): { start_z: numbe
 	const offset = HALF_D + height / HALF_DIVISOR;
 	return { start_z: offset, end_z: -offset };
 }
+
+export function advance_scroll(
+	current_z: number,
+	delta: number,
+	start_z: number,
+	end_z: number
+): number {
+	const next_z = current_z - CREDITS_SCROLL_SPEED * delta;
+	return next_z < end_z ? start_z : next_z;
+}
