+++ src/lib/game/player/camera-shake.svelte.ts
