+++ src/lib/simon/SimonBoard.svelte
@@ -4,7 +4,7 @@
 	import { fonts } from '$lib/game/fonts';
 	import { simon_board_input } from './simon-board-input';
 	import { BOARD_Y, BOARD_Z } from './board-config';
-	import type { ButtonColor, SimonBoardData } from './board-types';
+	import type { ButtonColor, SimonBoardData } from './types';
 
 	const INNER_RADIUS = 0.3;
 	const OUTER_RADIUS = 0.7;
