+++ src/lib/simon/score-display-types.ts
