--- src/lib/vitest-examples/Welcome.svelte
@@ -1,8 +1,0 @@
-<script>
-	import { greet } from './greet';
-
-	let { host = 'SvelteKit', guest = 'Vitest' } = $props();
-</script>
-
-<h1>{greet(host)}</h1>
-<p>{greet(guest)}</p>
