--- src/lib/vitest-examples/greet.spec.ts
@@ -1,8 +1,0 @@
-import { describe, it, expect } from 'vitest';
-import { greet } from './greet';
-
-describe('greet', () => {
-	it('returns a greeting', () => {
-		expect(greet('Svelte')).toBe('Hello, Svelte!');
-	});
-});
